+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the tool_vault plugin will be documented in this file.
 
+## [Unreleased]
+### Fixed
+- Restore plugins pre-check no longer fails on version mismatches for plugins listed in the `restorepreserveplugins` setting, since those plugins are skipped during restore.
+
 ## [4.5.0] - 2026-04-18
 ### Changed
 - Bumped minimum supported Moodle version to 4.5. Older Moodle versions (pre-4.5)